Move files to core module
