+++ core/shaper.js
