+++ index.html
@@ -21,7 +21,7 @@
     </div>
 
     <input class="left form-control form-control-sm round-right" style="width: 50px; margin-left: 40px;" type="number"
-      id="radius" name="radius" min="1" max="12" value="5" />
+      id="radius" name="radius" min="1" max="12" value="4" />
     <input type="button" value="Select Radius" class="left btn btn-sm btn-success round-left"
       id="select-radius"></input>
 
@@ -33,8 +33,8 @@
 
     <select id="cell-shape-option" class="left form-select form-select-sm round-right"
       style="width: 105px; margin-left: 10px;">
-      <option>Square</option>
       <option>Hexagon</option>
+      <option>Square</option>
     </select>
     <button class="btn btn-sm btn-success left round-left"><a id="cell-shape" href="#">Select Shape</button>
